Avoid "warning c4334: '<<': result of 32-bit shift implicitly converted to 64 bits (was 64-bit shift intended?)" from microsoft visual c.
Copied from Perforce
 Change: 194586

+++ mps/test/function/233.c
@@ -15,7 +15,7 @@ static void test(void)
 {
   size_t i;
   for (i = 0; i < 20; ++i) {
-    size_t unitSize = 1 << i;
+    size_t unitSize = (size_t)1 << i;
     mps_arena_t arena;
     mps_pool_t pool;
     mps_addr_t p;
